// client/src/pages/Profile/Profile.style.tsx
import styled from 'styled-components/macro';

export const Profile = styled.section`
  width: 100vw;
  height: auto;
  min-height: 720px;

  display: flex;
  flex-direction: column;
  align-items: center;
  justify-content: flex-start;
`;

export const TableWrapper = styled.div`
  padding-top: 40px;
  min-width: 800px;
  width: 60%;
  font-size: small;
`;